+++ changelog.md
@@ -3,3 +3,4 @@ vX.X.X
 
 - Automatic scan for COM ports performed on startup of app, closes #72.
 - Added ability to have multiple terminals open within the same NinjaTerm application window, closes #73.
+- "Busy COM port" error is now handled correctly, closes #74.